minor bugfixes

+++ benchmarks/basicdiscio.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
-dd bs=1M count=1024 if=/dev/zero of=/tmp/test
+
+#parsing arguments
+if [ $# -ne 1 ]; then
+	echo "Usage: $0 \"test directory\""
+	exit 0
+fi
+
+dd bs=1M count=10k if=/dev/zero of="$1"
 sleep 3
-dd bs=1M count=1024 if=/tmp/test of=/dev/null
+dd bs=1M count=10k if="$1" of=/dev/null
 rm /tmp/test
